chatbot/data/discord/requirements.txt
tensorflow-gpu
fire
regex
botframework-connector
toposort
tqdm